src/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
certifi==2022.6.15; python_version >= '3.6'
charset-normalizer==2.1.0; python_version >= '3.6'
colorama==0.4.5; platform_system == 'Windows'
decorator==4.4.2
idna==3.3; python_version >= '3.5'
imageio-ffmpeg==0.4.7; python_version >= '3.4'
imageio==2.19.3; python_version >= '3.4'
instagrapi==1.16.21
moviepy==1.0.3
numpy==1.23.1; python_version != '2.7'
pillow==9.2.0; python_version >= '3.7'
proglog==0.1.10
psycopg2-binary==2.9.3
pycryptodomex==3.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pydantic==1.9.1; python_full_version >= '3.6.1'
pysocks==1.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.20.0
requests==2.28.1
schedule==1.1.0
tqdm==4.64.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==4.3.0; python_version >= '3.7'
urllib3==1.26.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'